crew/pyproject.toml
[project]
name = "crew-dev-agents"
version = "0.1.0"
description = "Self-improving CrewAI developer agent team for open-source repo maintenance"
requires-python = ">=3.12"
dependencies = [
    "crewai[tools]>=0.152.0",
    "boto3>=1.35.0",
    "PyGithub>=2.5.0",
    "pydantic>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
